[feat] default Method를 부르기 위한 custom hook 작성
