+++ FE/src/pages/Calendar/index.tsx
@@ -4,22 +4,11 @@ import Modal from '../../components/PaymentMethod/Modal';
 import MenuBar from '../../components/Common/MenuBar';
 import Calendar from '../../components/Calendar';
 import styles from './calendar.module.scss';
-import { useRootData } from '../../store/PaymentMethod/paymentMethodHook';
+import useDefaultPayment from '../../service/useDefaultPayment';
 
 export default function CalendarPage() {
   const [modal, setModal] = useState(false);
-  const [defaultMethod, setDefaultMethod] = useState([]);
-
-  const storeData = useRootData(store => store.defaultMethods);
-
-  const getDefaultMethod = async () => {
-    const datas = await storeData;
-    setDefaultMethod(datas);
-  };
-
-  useEffect(() => {
-    getDefaultMethod();
-  }, []);
+  const defaultMethod = useDefaultPayment();
 
   return (
     <div className={styles.wrapper}>
